+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - customization demo ([#1207](https://github.com/catalyst-team/catalyst/pull/1207))
 - FAQ docs: multiple input and output keys, engine tutorial ([#1202](https://github.com/catalyst-team/catalyst/pull/1202))
 - minimal Config API example ([#1215](https://github.com/catalyst-team/catalyst/pull/1215))
+- SklearnCallback as integration of sklearn metrics ([#1198](https://github.com/catalyst-team/catalyst/pull/1198))
 
 ### Changed