+++ CHANGELOG.md
@@ -12,12 +12,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - customization demo ([#1207](https://github.com/catalyst-team/catalyst/pull/1207))
 - FAQ docs: multiple input and output keys, engine tutorial ([#1202](https://github.com/catalyst-team/catalyst/pull/1202))
 - minimal Config API example ([#1215](https://github.com/catalyst-team/catalyst/pull/1215))
+- Distributed RL example (Catalyst.RL 2.0 concepts) ([#1224](https://github.com/catalyst-team/catalyst/pull/1224))
 - SklearnCallback as integration of sklearn metrics ([#1198](https://github.com/catalyst-team/catalyst/pull/1198))
 
 ### Changed
 
 - tests moved to `tests` folder ([#1208](https://github.com/catalyst-team/catalyst/pull/1208))
 - pipeline tests moved to `tests/pipelines` ([#1215](https://github.com/catalyst-team/catalyst/pull/1215))
+- updated NeptuneLogger docstrings ([#1223](https://github.com/catalyst-team/catalyst/pull/1223))
 
 ### Removed