+++ CHANGELOG.md
@@ -29,7 +29,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - customizing what happens in `train()` notebook ([#1203](https://github.com/catalyst-team/catalyst/pull/1203))
 - transforms imports under catalyst.data ([#1211](https://github.com/catalyst-team/catalyst/pull/1211))
 - change layerwise to layerwise_params ([#1210](https://github.com/catalyst-team/catalyst/pull/1210))
-- Add torch metrics support ([#1195](https://github.com/catalyst-team/catalyst/issues/1195))
+- add torch metrics support ([#1195](https://github.com/catalyst-team/catalyst/issues/1195))
+- add Config API support for BatchTransformCallback ([#1209](https://github.com/catalyst-team/catalyst/issues/1209))
 
 
 ## [21.04.2] - 2021-04-30